pyproject.toml
[project]
name = "thesistand"
version = "0.1.0"
description = "thesistand Projects"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.128.7",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.8",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "llama-parse>=0.6.92",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]



[project.optional-dependencies]

test = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

dev = [
    "ruff>=0.14.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
